+++ source/rwg/assemblers/indexing.hpp
@@ -12,30 +12,103 @@
 
 /**
 * @file
-* Index generators for assemblers.
+* Index sets and generators for assemblers.
 */
 
-#ifndef ASSEMBLERS_INDGEN_H
-#define ASSEMBLERS_INDGEN_H
+#ifndef ASSEMBLERS_IND_H
+#define ASSEMBLERS_IND_H
 
 #include <vector>
 
 #include "types.hpp"
+#include "geometry/mesh/triangle_mesh.hpp"
+#include "rwg/function_space.hpp"
 
 
 namespace bem
 {
 
-// Forward declarations
-template <uint8_t dim> class TriangleMesh;
-
 /**
-* \defgroup assmutil Utility
 * \ingroup assm
-* @brief Utility and helper functions for assemblers.
 * @{
 */
 
+/**
+* @brief Data structure of index sets defining a block of a matrix.
+*/
+struct IndexSet
+{
+public:
+
+    /**
+    * @brief Constructs an `IndexSet` object with row and column index lists.
+    * @param[in] rows - List of row indices.
+    * @param[in] cols - List of column indices.
+    */
+    IndexSet(
+        ConstEigRef<EigRowVec<Index>> rows,
+        ConstEigRef<EigRowVec<Index>> cols
+        ): rows_(rows), cols_(cols) {};
+
+
+    /**
+    * @brief Constructs an `IndexSet` object for a contiguous block.
+    * @param[in] row_start - Starting row index.
+    * @param[in] col_start - Starting column index.
+    * @param[in] num_rows - Number of rows in the block.
+    * @param[in] num_cols - Number of columns in the block.
+    */
+    IndexSet(
+        const Index row_start,
+        const Index col_start,
+        const Index num_rows,
+        const Index num_cols
+        ):
+        IndexSet(
+            EigRowVec<Index>::LinSpaced(num_rows, row_start, row_start + num_rows - 1),
+            EigRowVec<Index>::LinSpaced(num_cols, col_start, col_start + num_cols - 1)
+            ) {};
+
+
+    /**
+    * @brief Returns a read-only reference to the row indices.
+    * @return Read-only reference to row indices.
+    */
+    const EigRowVec<Index>& rows() const { return rows_; };
+
+
+    /**
+    * @brief Returns a read-only reference to the column indices.
+    * @return Read-only reference to column indices.
+    */
+    const EigRowVec<Index>& cols() const { return cols_; };
+
+
+    /**
+    * @brief Returns the number of rows.
+    * @return Number of rows.
+    */
+    Index num_rows() const { return rows_.size(); };
+
+
+    /**
+    * @brief Returns the number of columns.
+    * @return Number of columns.
+    */
+    Index num_cols() const { return cols_.size(); };
+
+
+protected:
+
+    const EigRowVec<Index> rows_;
+    const EigRowVec<Index> cols_;
+
+};
+
+
+/**
+* @brief Class to generate index sets and maps across element types.
+*/
 class IndexGenerator
 {
 public:
@@ -55,15 +128,6 @@ public:
 
 
     /**
-     * @brief Generates all possible pairs of triangle indices for a given triangle mesh.
-     * @param[in] mesh - Triangle mesh.
-     * @return All possible triangle index pairs, with observation indices in the first row,
-     * and source indices in the second row.
-     */
-    static EigMatNX<Index, 2> elem_pairs(const TriangleMesh<3>& mesh);
-
-
-    /**
     * @brief Makes all possible pairs of triangle indices for given observation and source triangle indices.
     * @param[in] obs_elems - Observation triangle indices.
     * @param[in] src_elems - Source triangle indices.
@@ -77,33 +141,35 @@ public:
 
 
     /**
-    * @brief Makes all possible pairs of given triangle indices.
-    * @param[in] elems - Triangle indices.
-    * @return Triangle index pairs, with observation indices in the first row,
-    * and source indices in the second row.
-    */
-    static EigMatNX<Index, 2> elem_pairs(ConstEigRef<EigRowVec<Index>> elems);
-
-
-    /**
-    * @brief Makes self-pairs of given triangle indices.
-    * @param[in] elems - Triangle indices.
-    * @return Triangle index pairs, with observation indices in the first row,
-    * and source indices in the second row.
-    */
-    static EigMatNX<Index, 2> elem_pairs_self(ConstEigRef<EigRowVec<Index>> elems);
-
-
-    /**
-    * @brief Returns all unique element indices associated with given edge indices.
+    * @brief Returns all unique element pairs associated with a given index set.
     * @param[in] mesh - Triangle mesh.
-    * @param[in] edges - Edge indices.
-    * @return Triangle indices.
+    * @param[in] index_set - Index set defining degrees of freedom.
+    * @param[in] row_dof - Degrees of freedom associated with rows.
+    * @param[in] col_dof - Degrees of freedom associated with columns.
+    * @return Triangle index pairs.
     */
-    static EigRowVec<Index> elems_from_edges(
+    static EigMatNX<Index, 2> elem_pairs(
         const TriangleMesh<3>& mesh,
-        ConstEigRef<EigRowVec<Index>> edges
+        const IndexSet& index_set,
+        const OperatorDof row_dof,
+        const OperatorDof col_dof
         );
+
+
+    /**
+    * @brief Returns unique pairs from given pairs.
+    * @param[in] pairs - Index pairs.
+    * @return Unique index pairs.
+    */
+    static EigMatNX<Index, 2> unique_pairs(
+        ConstEigRef<EigMatNX<Index, 2>> pairs
+        );
+
+
+
+
+
+
 
 
     /**
@@ -164,13 +230,17 @@ public:
         );
 
 
+protected:
+
     /**
-    * @brief Returns unique pairs from given pairs.
-    * @param[in] pairs - Index pairs.
-    * @return Unique index pairs.
+    * @brief Returns all unique element indices associated with given edge indices.
+    * @param[in] mesh - Triangle mesh.
+    * @param[in] edges - Edge indices.
+    * @return Triangle indices.
     */
-    static EigMatNX<Index, 2> unique_pairs(
-        ConstEigRef<EigMatNX<Index, 2>> pairs
+    static EigRowVec<Index> elems_from_edges(
+        const TriangleMesh<3>& mesh,
+        ConstEigRef<EigRowVec<Index>> edges
         );
 
 };
@@ -182,7 +252,7 @@ public:
 }
 
 #ifndef BEM_LINKED
-#include "rwg/assemblers/index_generator.cpp"
+#include "rwg/assemblers/indexing.cpp"
 #endif
 
 #endif