+++ source/rwg/operators/base.hpp
@@ -43,17 +43,17 @@ class OperatorBase
 public:
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the testing function space.
-    * @return Number of observation degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the testing function space.
+    * @return Observation degrees of freedom.
     */
-    virtual uint8_t obs_dof() const = 0;
+    virtual OperatorDof obs_dof() const = 0;
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the expansion function space.
-    * @return Number of source degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the expansion function space.
+    * @return Source degrees of freedom.
     */
-    virtual uint8_t src_dof() const = 0;
+    virtual OperatorDof src_dof() const = 0;
 
 
     /**