+++ source/rwg/operators/gram.hpp
@@ -70,17 +70,17 @@ public:
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the testing function space.
-    * @return Number of observation degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the testing function space.
+    * @return Observation degrees of freedom.
     */
-    uint8_t obs_dof() const override { return 3; };
+    OperatorDof obs_dof() const override { return OperatorDof::EDGE; };
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the expansion function space.
-    * @return Number of source degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the expansion function space.
+    * @return Source degrees of freedom.
     */
-    uint8_t src_dof() const override { return 3; };
+    OperatorDof src_dof() const override { return OperatorDof::EDGE; };
 
 
     /**
@@ -165,17 +165,17 @@ public:
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the testing function space.
-    * @return Number of observation degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the testing function space.
+    * @return Observation degrees of freedom.
     */
-    uint8_t obs_dof() const override { return 3; };
+    OperatorDof obs_dof() const override { return OperatorDof::EDGE; };
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the expansion function space.
-    * @return Number of source degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the expansion function space.
+    * @return Source degrees of freedom.
     */
-    uint8_t src_dof() const override { return 3; };
+    OperatorDof src_dof() const override { return OperatorDof::EDGE; };
 
 
     /**