+++ source/rwg/projectors/single_layer.hpp
@@ -57,10 +57,10 @@ public:
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the expansion function space.
-    * @return Number of source degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the expansion function space.
+    * @return Source degrees of freedom.
     */
-    uint8_t src_dof() const override { return 3; };
+    OperatorDof src_dof() const override { return OperatorDof::EDGE; };
 
 
     /**
@@ -119,10 +119,10 @@ public:
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the expansion function space.
-    * @return Number of source degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the expansion function space.
+    * @return Source degrees of freedom.
     */
-    uint8_t src_dof() const override { return 1; };
+    OperatorDof src_dof() const override { return OperatorDof::FACE; };
 
 
     /**
@@ -177,10 +177,10 @@ public:
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the expansion function space.
-    * @return Number of source degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the expansion function space.
+    * @return Source degrees of freedom.
     */
-    uint8_t src_dof() const override { return 1; };
+    OperatorDof src_dof() const override { return OperatorDof::FACE; };
 
 
     /**
@@ -238,10 +238,10 @@ public:
 
 
     /**
-    * @brief Returns the number of degrees of freedom per triangle for the expansion function space.
-    * @return Number of source degrees of freedom per triangle.
+    * @brief Returns the degrees of freedom for the expansion function space.
+    * @return Source degrees of freedom.
     */
-    uint8_t src_dof() const override { return 3; };
+    OperatorDof src_dof() const override { return OperatorDof::EDGE; };
 
 
     /**